+++ lib/vespa_ruby/yql_query.rb
@@ -2,7 +2,7 @@
 
 # rbs_inline: enabled
 
-require "active_support"
+require "active_support/core_ext/object/blank"
 
 module VespaRuby
   class YqlQuery